streamlit-hello-world-app/requirements.txt
streamlit
pandas
databricks-sdk